ieeeFlags to force dependency on an expression
A new version of ieeeFlags() is added that takes an argument which is an
expression that must be evaluated before the ieeeFlags are read.  This
is needed because D does not have something like c99 FENV_ACCESS pragma,
and so there is no guarentee that ieeeFlags will be accessed after the
expression of interest.